refactor for simplicity
