+++ stagehand/src/utils.ts
@@ -64,4 +64,29 @@ export function jsonSchemaToZod(schema: any): ZodTypeAny {
     default:
       return z.any();
   }
+}
+
+/**
+ * Sanitizes a message to ensure it's properly formatted JSON
+ * @param message The message to sanitize
+ * @returns A sanitized JSON string
+ */
+export function sanitizeMessage(message: any): string {
+  try {
+    // Ensure the message is properly stringified JSON
+    if (typeof message === 'string') {
+      JSON.parse(message); // Validate JSON structure
+      return message;
+    }
+    return JSON.stringify(message);
+  } catch (error) {
+    return JSON.stringify({
+      jsonrpc: '2.0',
+      error: {
+        code: -32700,
+        message: 'Parse error',
+      },
+      id: null,
+    });
+  }
 }
